+++ src/main/java/com/nhnacademy/store99/bookstore/book_tag/service/BookTagService.java
@@ -1,6 +1,7 @@
 package com.nhnacademy.store99.bookstore.book_tag.service;
 
-import com.nhnacademy.store99.bookstore.book.response.BookResponse;
+
+import com.nhnacademy.store99.bookstore.book.dto.response.BookResponse;
 import java.util.List;
 
 public interface BookTagService {
